Add video example, remove helper model
