+++ examples/ImageClassifier-video/index.html
@@ -4,13 +4,13 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Image Classifier Example</title>
+    <title>Webcam Image Classifier Example</title>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.7.0/p5.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.9.0/addons/p5.dom.min.js"></script>
     <script src="../../dist/ml5.js"></script>
   </head>
   <body>
-    <h1>Image classification using MobileNet</h1>
+    <h1>Webcam Image classification using MobileNet</h1>
     <script src="sketch.js"></script>
   </body>
 </html>
