--- examples/xImageClassifier/index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Image Classifier Example</title>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.7.0/p5.js"></script>
-    <script src="../../dist/ml5.js"></script>
-  </head>
-  <body>
-    <h1>Image classification using MobileNet</h1>
-    <p>The MobileNet model labeled this as <span id="result">...</span> with a confidence of <span id="probability">...</span>.</p>
-    <img src="images/bird.jpg" id="image" width="400"/>
-    <video src="images/flower.webm" id="vid" width="250"/>
-    <!-- <video controls width="250">
-      <source src="images/flower.webm" id="vid"/>
-        
-      Download the
-      <a href="images/flower.webm">WEBM</a>
-      video.
-    </video> -->
-    <script src="script.js"></script>
-  </body>
-</html>
